+++ Core/Base/AjaxForms/SalesController.php
@@ -144,6 +144,15 @@ abstract class SalesController extends PanelController
     {
         $this->setTemplate(false);
 
+        // comprobamos los permisos
+        if (false === $this->permissions->allowDelete) {
+            self::toolBox()::i18nLog()->warning('not-allowed-delete');
+            $this->response->setContent(
+                json_encode(['ok' => false, 'messages' => self::toolBox()::log()::read('', $this->logLevels)])
+            );
+            return false;
+        }
+
         $model = $this->getModel();
         if (false === $model->delete()) {
             $this->response->setContent(
@@ -343,6 +352,16 @@ abstract class SalesController extends PanelController
     protected function saveDocAction(): bool
     {
         $this->setTemplate(false);
+
+        // comprobamos los permisos
+        if (false === $this->permissions->allowUpdate) {
+            self::toolBox()::i18nLog()->warning('not-allowed-modify');
+            $this->response->setContent(
+                json_encode(['ok' => false, 'messages' => self::toolBox()::log()::read('', $this->logLevels)])
+            );
+            return false;
+        }
+
         $this->dataBase->beginTransaction();
 
         $model = $this->getModel();
@@ -392,6 +411,15 @@ abstract class SalesController extends PanelController
     {
         $this->setTemplate(false);
 
+        // comprobamos los permisos
+        if (false === $this->permissions->allowUpdate) {
+            self::toolBox()::i18nLog()->warning('not-allowed-modify');
+            $this->response->setContent(
+                json_encode(['ok' => false, 'messages' => self::toolBox()::log()::read('', $this->logLevels)])
+            );
+            return false;
+        }
+
         $model = $this->getModel();
         $receipts = $model->getReceipts();
         if (empty($receipts)) {
@@ -416,6 +444,16 @@ abstract class SalesController extends PanelController
     protected function saveStatusAction(): bool
     {
         $this->setTemplate(false);
+
+        // comprobamos los permisos
+        if (false === $this->permissions->allowUpdate) {
+            self::toolBox()::i18nLog()->warning('not-allowed-modify');
+            $this->response->setContent(
+                json_encode(['ok' => false, 'messages' => self::toolBox()::log()::read('', $this->logLevels)])
+            );
+            return false;
+        }
+
         if ($this->getModel()->editable && false === $this->saveDocAction()) {
             return false;
         }
